+++ src/Filters/BellFilter.h
@@ -15,7 +15,7 @@ namespace Raycer
 
 		BellFilter();
 
-		double getWeightX(double x) override;
-		double getWeightY(double y) override;
+		float getWeightX(float x) override;
+		float getWeightY(float y) override;
 	};
 }
